Factored out dag_file.[ch]
